requirements.txt
chromadb
sentence-transformers
numpy
scipy
pandas
scikit-learn
tqdm
langchain
faiss-cpu
nltk